llm-code-deployment/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
PyGithub
python-jose[cryptography]
passlib[bcrypt]
openai
langchain
langchain-openai
beautifulsoup4
gunicorn
loguru
pydantic-settings
uvloop
httptools
pytest
httpx